feat: a magical web ui (#10)
packages visualised as moonlit ascii-trees that dance in the gloaming.

![what-packages-do-in-dark](https://user-images.githubusercontent.com/58871/106445605-220b8c00-6477-11eb-834c-793126a5df3c.gif)

more prosaically, it adds a web ui that scatters your list of IPFS backed, forest hosted packages across the page, each one represented by a ╪ that links to the npm page for that package.

the trees are distributed by their package name. locality is preserved such that packages with similar names will huddle together.

It is an HTML poem.